Move custom gitignored work into customgitignored package
